Use parallel gem for parallel calculation of training data files
This also changes the creation process to a per feature workflow, where
all edits for a certain feature are calculated, written to an arff file
and only then the next feature is processed.

Having this workflow allows interrupting the building process after each
feature without loosing the data for already calculated features.